+++ pytest/test_bem_hierarchical_toeplitz_matrices.py
@@ -16,15 +16,14 @@ from capytaine.bodies.predefined.spheres import Sphere
 from capytaine.bodies.predefined.cylinders import HorizontalCylinder, VerticalCylinder
 
 from capytaine.bem.problems_and_results import RadiationProblem
-from capytaine.bem.solver import Nemoh
 from capytaine.io.xarray import assemble_dataset
 
 from capytaine.meshes.geometry import xOz_Plane, yOz_Plane
 
 from capytaine.matrices.low_rank import LowRankMatrix
 
-solver_with_sym = Nemoh(hierarchical_matrices=True, ACA_distance=8, matrix_cache_size=0)
-solver_without_sym = Nemoh(hierarchical_matrices=False, ACA_distance=8, matrix_cache_size=0)
+solver_with_sym = cpt.BEMSolver(engine=cpt.HierarchicalToeplitzMatrixEngine(ACA_distance=8, matrix_cache_size=0))
+solver_without_sym = cpt.BEMSolver(engine=cpt.BasicMatrixEngine(matrix_cache_size=0))
 # Use a single solver in the whole module to avoid reinitialisation of the solver (0.5 second).
 # Do not use a matrix cache in order not to risk influencing a test with another.
 
@@ -158,7 +157,7 @@ def test_low_rank_matrices():
     two_distant_buoys = FloatingBody.join_bodies(buoy, buoy.translated_x(20))
     two_distant_buoys.mesh._meshes[1].name = "other_buoy_mesh"
 
-    S, V = solver_with_sym.build_matrices(two_distant_buoys.mesh, two_distant_buoys.mesh, 0.0, -np.infty, 1.0)
+    S, V = solver_with_sym.engine.build_matrices(two_distant_buoys.mesh, two_distant_buoys.mesh, 0.0, -np.infty, 1.0, solver_with_sym.green_function)
     assert isinstance(S.all_blocks[0, 1], LowRankMatrix)
     assert isinstance(S.all_blocks[1, 0], LowRankMatrix)
     # S.plot_shape()
@@ -199,12 +198,8 @@ def test_array_of_spheres():
     #
     array = buoy.assemble_regular_array(distance=4.0, nb_bodies=(3, 1))
 
-    settings = dict(cache_rankine_matrices=False, matrix_cache_size=0)
-    nemoh_without_sym = Nemoh(hierarchical_matrices=False, **settings)
-    nemoh_with_sym = Nemoh(hierarchical_matrices=True, **settings)
-
-    fullS, fullV = nemoh_without_sym.build_matrices(array.mesh, array.mesh, 0.0, -np.infty, 1.0)
-    S, V = nemoh_with_sym.build_matrices(array.mesh, array.mesh, 0.0, -np.infty, 1.0)
+    fullS, fullV = solver_without_sym.engine.build_matrices(array.mesh, array.mesh, 0.0, -np.infty, 1.0, solver_without_sym.green_function)
+    S, V = solver_with_sym.engine.build_matrices(array.mesh, array.mesh, 0.0, -np.infty, 1.0, solver_with_sym.green_function)
 
     assert isinstance(S, cpt.matrices.block.BlockMatrix)
     assert np.allclose(S.full_matrix(), fullS)
@@ -212,8 +207,8 @@ def test_array_of_spheres():
 
     problem = RadiationProblem(body=array, omega=1.0, radiating_dof="2_0__Heave", sea_bottom=-np.infty)
 
-    result = nemoh_with_sym.solve(problem)
-    result2 = nemoh_without_sym.solve(problem)
+    result = solver_with_sym.solve(problem)
+    result2 = solver_without_sym.solve(problem)
 
     assert np.isclose(result.added_masses['2_0__Heave'], result2.added_masses['2_0__Heave'], atol=15.0)
     assert np.isclose(result.radiation_dampings['2_0__Heave'], result2.radiation_dampings['2_0__Heave'], atol=15.0)