+++ pytest/test_bem_problems_and_results.py
@@ -20,7 +20,6 @@ from capytaine.bodies.predefined.cylinders import HorizontalCylinder
 
 from capytaine.bem.problems_and_results import LinearPotentialFlowProblem, DiffractionProblem, RadiationProblem, \
     LinearPotentialFlowResult, DiffractionResult, RadiationResult
-from capytaine.bem.solver import Nemoh
 from capytaine.io.xarray import problems_from_dataset, assemble_dataset
 
 from capytaine.io.legacy import import_cal_file